по поводу коментариев в коде, если так не нужно, значит не буду оставлять, это не проблема. развел в разные классы но потом начались проблеммы, не компилировалась программа. решением стало gradlew clean, о сначала и саму gradlew пришлось настроить, и теперь все работает
