+++ src/main/java/school/faang/bjs2_79362/Character.java
@@ -12,7 +12,6 @@ public abstract class Character {
 
     private static final int DEFAULT_HEALTH = 100;
 
-    // Конструктор с базовыми характеристиками
     public Character(String name) {
         this.name = name;
         this.strength = 5;
@@ -21,7 +20,6 @@ public abstract class Character {
         this.health = DEFAULT_HEALTH;
     }
 
-    // Конструктор с заданными характеристиками
     public Character(String name, int strength, int agility, int intelligence) {
         this.name = name;
         this.strength = strength;
@@ -30,48 +28,12 @@ public abstract class Character {
         this.health = DEFAULT_HEALTH;
     }
 
-    // Абстрактный метод атаки
     public abstract void attack(Character opponent);
 
-    // Метод уменьшения здоровья (не может быть меньше 0)
     public void reduceHealth(int damage) {
         if (damage < 0) {
-            throw new IllegalArgumentException("Урон не может быть отрицательным");
+            throw new IllegalArgumentException("Урон не может быть отрицательным!");
         }
         this.health = Math.max(0, this.health - damage);
-    }
-
-    //  Класс Archer
-    public static class Archer extends Character {
-        public Archer(String name) {
-            super(name);
-            this.strength = 3;
-            this.agility = 10;
-            this.intelligence = 5;
-        }
-
-        @Override
-        public void attack(Character opponent) {
-            System.out.println(name + " атакует " + opponent.getName() +
-                    " с силой удара равной ловкости " + agility);
-            opponent.reduceHealth(agility);
-        }
-    }
-
-    //  Класс Warrior
-    public static class Warrior extends Character {
-        public Warrior(String name) {
-            super(name);
-            this.strength = 10;
-            this.agility = 5;
-            this.intelligence = 3;
-        }
-
-        @Override
-        public void attack(Character opponent) {
-            System.out.println(name + " атакует " + opponent.getName() +
-                    " с силой " + strength);
-            opponent.reduceHealth(strength);
-        }
     }
 }
